requirements.txt
asgiref>=3.7

flask-restx>=1.3.0
Werkzeug>=3.0

requests==2.32.5
aiohttp==3.13.3
aiosqlite==0.22.1

python-slugify==8.0.4

aiosmtplib==5.0.0
python-dotenv==1.1.1

hypercorn==0.18.0
